refactor: simplify entities and reorganize (#6)
