+++ app/apis/catalog/hca-atlas-tracker/common/utils.ts
@@ -10,5 +10,5 @@ export function getAtlasId(atlas: HCAAtlasTrackerAtlas): string {
 export function getComponentAtlasId(
   componentAtlas: HCAAtlasTrackerComponentAtlas
 ): string {
-  return componentAtlas.cxgId;
+  return componentAtlas.cxgDatasetId;
 }
