+++ files/constants.ts
@@ -1,0 +1,9 @@
+export const DATA_SOURCE = {
+  defaultListParams: {
+    size: "100",
+  },
+  defaultParams: {
+    catalog: "dcp34",
+  },
+  url: "https://service.azul.data.humancellatlas.org/index",
+};
